Move from anyvalue implementation
